refactor elasticsearch configuration
