+++ dev/elasticsearch.yml
@@ -9,6 +9,9 @@ cluster.name: o2rcluster
 
 network.host: 0.0.0.0
 
+xpack.security.enabled: false
+xpack.monitoring.enabled: false
+
 # this value is required because we set "network.host"
 # be sure to modify it appropriately for a production cluster deployment
 discovery.zen.minimum_master_nodes: 1
