+++ package.json
@@ -31,5 +31,12 @@
     "node-elasticsearch-sync": "github:nuest/node-elasticsearch-sync",
     "passport": "^0.4.0",
     "response-time": "2.3.2"
+  },
+  "devDependencies": {
+    "chai": "^4.1.2",
+    "chai-datetime": "^1.5.0",
+    "mocha": "^3.5.0",
+    "mongojs": "^2.4.1",
+    "request": "^2.83.0"
   }
 }
